Rename new shardmanager service to shard-distributor
