+++ common/service/metrics.go
@@ -36,8 +36,8 @@ func GetMetricsServiceIdx(serviceName string, logger log.Logger) metrics.Service
 		return metrics.Matching
 	case Worker:
 		return metrics.Worker
-	case ShardManager:
-		return metrics.ShardManager
+	case ShardDistributor:
+		return metrics.ShardDistributor
 	default:
 		logger.Fatal("Unknown service name for metrics!")
 	}
