+++ service/sharddistributor/service_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package shardmanager
+package sharddistributor
 
 import (
 	"sync/atomic"
